Addressing review comment
